# Optimize Solidity Code by Reducing Redundant Calls, Consolidating Checks, and Adjusting Variable Types in Gas.sol