--- a/foundry.toml
+++ b/foundry.toml
@@ -3,7 +3,7 @@ src = 'src'
 out = 'out'
 libs = ['lib']
 optimizer = true
-optimizer_runs = 100
+optimizer_runs = 10
 via_ir = false
 
 # See more config options https://github.com/foundry-rs/foundry/tree/master/config