This pull request introduces several optimizations to the Solidity code in the Gas.sol file. The main changes include:

1. Consolidating conditional checks: In several functions and modifiers, multiple conditional checks have been consolidated into single require statements. This makes the code more concise and reduces the number of require statements.
2. Reducing redundant function calls: In the onlyAdminOrOwner modifier, the checkForAdmin function was being called twice. This has been optimized to call the function only once, reducing gas costs.
3. Optimizing variable types: In the ImportantStruct struct, the valueA and valueB fields were defined as uint256, but the comments indicated that these values would only have a maximum of 3 digits. These fields have been changed to uint16 to save on gas costs.
4. Removing unnecessary Ether transfers: The receive and fallback functions were implemented to immediately return any received Ether back to the sender. This is generally inappropriate and wastes gas. These functions have been left empty to allow the contract to receive Ether without immediately transferring it back.

How much reduce Deployment gas Cost?
This change reduce gas cost from 2412246 to 1917933 in my local environment.

## Before 
```
❯ forge test --gas-report        
[⠢] Compiling...
[⠔] Compiling 20 files with 0.8.0
[⠒] Solc 0.8.0 finished in 3.33s
Compiler run successful!

Ran 16 tests for test/Gas.UnitTests.t.sol:GasTest
[PASS] testAddHistory() (gas: 277)
[PASS] testAddToWhitelist(address,uint256) (runs: 256, μ: 26072, ~: 26075)
[PASS] testBalanceOf() (gas: 12267)
[PASS] testCheckForAdmin() (gas: 19614)
[PASS] testGetPaymentHistory() (gas: 189)
[PASS] testGetPaymentStatus(address) (runs: 256, μ: 321, ~: 321)
[PASS] testGetTradingMode() (gas: 276)
[PASS] testTransfer(uint256,address) (runs: 256, μ: 209228, ~: 212057)
[PASS] testWhiteTranferAmountUpdate(address,address,uint256,string,uint256) (runs: 256, μ: 361297, ~: 361559)
[PASS] test_admins() (gas: 36609)
[PASS] test_onlyOwner(address,uint256) (runs: 256, μ: 27847, ~: 27961)
[PASS] test_tiers(address,uint256) (runs: 256, μ: 76186, ~: 76270)
[PASS] test_tiersReverts(address,uint256) (runs: 256, μ: 25802, ~: 25802)
[PASS] test_whitelistEvents(address,address,uint256,string,uint256) (runs: 256, μ: 350806, ~: 354094)
[PASS] test_whitelistEvents(address,uint256) (runs: 256, μ: 78251, ~: 78391)
[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 256, μ: 350396, ~: 354278)
Suite result: ok. 16 passed; 0 failed; 0 skipped; finished in 366.69ms (1.11s CPU time)
| src/Gas.sol:GasContract contract |                 |        |        |        |         |
|----------------------------------|-----------------|--------|--------|--------|---------|
| Deployment Cost                  | Deployment Size |        |        |        |         |
| 2412246                          | 12316           |        |        |        |         |
| Function Name                    | min             | avg    | median | max    | # calls |
| addToWhitelist                   | 13862           | 58510  | 85056  | 85128  | 8       |
| administrators                   | 2613            | 2613   | 2613   | 2613   | 5       |
| balanceOf                        | 637             | 2137   | 2637   | 2637   | 8       |
| balances                         | 642             | 1142   | 642    | 2642   | 4       |
| checkForAdmin                    | 12004           | 12004  | 12004  | 12004  | 1       |
| getPaymentStatus                 | 873             | 873    | 873    | 873    | 1       |
| transfer                         | 189105          | 190605 | 191105 | 191105 | 4       |
| whiteTransfer                    | 94999           | 96332  | 96999  | 96999  | 3       |
| whitelist                        | 619             | 619    | 619    | 619    | 2       |

Ran 1 test suite in 372.82ms (366.69ms CPU time): 16 tests passed, 0 failed, 0 skipped (16 total tests)
```
## After
```
❯ forge test --gas-report                                           
[⠢] Compiling...
No files changed, compilation skipped

Ran 16 tests for test/Gas.UnitTests.t.sol:GasTest
[PASS] testAddHistory() (gas: 386)
[PASS] testAddToWhitelist(address,uint256) (runs: 256, μ: 25877, ~: 25886)
[PASS] testBalanceOf() (gas: 12485)
[PASS] testCheckForAdmin() (gas: 20472)
[PASS] testGetPaymentHistory() (gas: 452)
[PASS] testGetPaymentStatus(address) (runs: 256, μ: 430, ~: 430)
[PASS] testGetTradingMode() (gas: 694)
[PASS] testTransfer(uint256,address) (runs: 256, μ: 206900, ~: 210229)
[PASS] testWhiteTranferAmountUpdate(address,address,uint256,string,uint256) (runs: 256, μ: 357824, ~: 358177)
[PASS] test_admins() (gas: 38763)
[PASS] test_onlyOwner(address,uint256) (runs: 256, μ: 27977, ~: 28081)
[PASS] test_tiers(address,uint256) (runs: 256, μ: 74747, ~: 74895)
[PASS] test_tiersReverts(address,uint256) (runs: 256, μ: 24084, ~: 24084)
[PASS] test_whitelistEvents(address,address,uint256,string,uint256) (runs: 256, μ: 345809, ~: 349902)
[PASS] test_whitelistEvents(address,uint256) (runs: 256, μ: 77043, ~: 77139)
[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 256, μ: 347066, ~: 350277)
Suite result: ok. 16 passed; 0 failed; 0 skipped; finished in 342.83ms (1.07s CPU time)
| src/Gas.sol:GasContract contract |                 |        |        |        |         |
|----------------------------------|-----------------|--------|--------|--------|---------|
| Deployment Cost                  | Deployment Size |        |        |        |         |
| 1917933                          | 9652            |        |        |        |         |
| Function Name                    | min             | avg    | median | max    | # calls |
| addToWhitelist                   | 12190           | 57209  | 83386  | 83386  | 8       |
| administrators                   | 3053            | 3053   | 3053   | 3053   | 5       |
| balanceOf                        | 746             | 2246   | 2746   | 2746   | 8       |
| balances                         | 596             | 1096   | 596    | 2596   | 4       |
| checkForAdmin                    | 12444           | 12444  | 12444  | 12444  | 1       |
| getPaymentStatus                 | 982             | 982    | 982    | 982    | 1       |
| transfer                         | 143823          | 177273 | 188423 | 188423 | 4       |
| whiteTransfer                    | 75029           | 76362  | 77029  | 77029  | 3       |
| whitelist                        | 904             | 904    | 904    | 904    | 2       |

Ran 1 test suite in 347.03ms (342.83ms CPU time): 16 tests passed, 0 failed, 0 skipped (16 total tests)

```